整理结构,调整窗口适配大小
