+++ demo/demo01/js/index.js
@@ -2,7 +2,6 @@
 "use strict";
 window.onload = function () {
   fullpage.init('#sectionContent', {
-    autoScroll: 3000,
     pageSpeed: 500,
     beforeLeave: function (leaveIndex, nowIndex) {
       if (nowIndex === 2) {
