+++ src/fullpage.css
@@ -1,17 +1,16 @@
-html,
-body {
+.fp-wrap{
   overflow: hidden;
   margin: 0;
   padding: 0;
-  width: 100%;
-  height: 100%;
   -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
 }
 
 .fp-section-content {
+  width: 100%;
+  height: 100%;
   margin: 0;
   padding: 0;
-  transition: all 0.5s ease-in-out;
+  /*transition: all 0.5s ease-in-out;*/
 }
 
 .fp-section-content .fp-section {
